+++ lib/primitives/takeTwo.test.ts
@@ -1,0 +1,35 @@
+import { assertEquals } from "@std/assert";
+import { any, many, take, takeTwo } from "../main.ts";
+
+// Explore a search space
+const oneOrTwoItems = any(take, takeTwo);
+const explore = many(oneOrTwoItems);
+
+Deno.test("explore", () => {
+  assertEquals(explore.parse("many"), {
+    success: true,
+    results: [
+      {
+        remaining: "",
+        value: ["m", "a", "n", "y"],
+        position: { line: 1, column: 4 },
+      },
+      {
+        remaining: "",
+        value: ["m", "a", "ny"],
+        position: { line: 1, column: 4 },
+      },
+      {
+        remaining: "",
+        value: ["m", "an", "y"],
+        position: { line: 1, column: 4 },
+      },
+      {
+        remaining: "",
+        value: ["ma", "n", "y"],
+        position: { line: 1, column: 4 },
+      },
+      { remaining: "", value: ["ma", "ny"], position: { line: 1, column: 4 } },
+    ],
+  });
+});
