+++ src/parser.test.ts
@@ -1,0 +1,24 @@
+import { assertEquals, assertIsError, assertThrows } from "@std/assert";
+import { and, literal, number, take } from "../lib/main.ts";
+import { ParseError } from "./errors.ts";
+
+const thrw = and([number, literal("then"), number]);
+
+Deno.test("parse error", () => {
+  assertEquals(take.parseOrThrow("monad"), "m");
+
+  assertThrows(() => (thrw.parseOrThrow("1 next 2")));
+
+  try {
+    thrw.parseOrThrow("1 next 2");
+  } catch (error: unknown) {
+    const errorMessage = `at line 1, column 2
+	1 next 2
+	  ^
+Reason: Expected 'then', but got 'next'`;
+    assertIsError(error, ParseError);
+    if (error instanceof ParseError) {
+      assertEquals(error.message, errorMessage);
+    }
+  }
+});
